Replaced size_t by unsigned int

+++ include/vml.h
@@ -6,7 +6,7 @@
 /*   By: gpinchon <gpinchon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/12 17:13:36 by gpinchon          #+#    #+#             */
-/*   Updated: 2016/11/23 23:43:29 by gpinchon         ###   ########.fr       */
+/*   Updated: 2016/11/24 00:34:33 by gpinchon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -273,6 +273,6 @@ VEC3			plane_normal(VEC3 position, PRIMITIVE p);
 VEC3			cone_normal(VEC3 position, PRIMITIVE p);
 
 char			float_equal(float a, float b);
-void			*vml_memset(void *dst, int c, size_t n);
+void			*vml_memset(void *dst, int c, unsigned int n);
 
 #endif
